reorganizes some code
